Error correction by cascade algorithm
In this algorithm, a random permutation will be firstly generated, and the secret key will be in this permutation. The new secret key will be divided into several blocks (typically the block size is ceil(0.73/ErrorRate)). Then, a parity check will be applied, and by comparing the parity, whether an error occurs can be determined. By the binary search, the exact location of the error can be known. After 4 passes like the above procedure, most of errors will be corrected.